requirements.txt
python-dotenv
pyparsing
networkx
matplotlib